# docs/swagger.yaml
definitions:
  model.LocalPatchBinaryResponse:
    type: object
  model.LocalPatchSourceCodeResponse:
    type: object
  model.PingCodeSimResponse:
    properties:
      full_text:
        type: string
    type: object
  model.SearchSourceCodeResponse:
    properties:
      project_files:
        items:
          $ref: '#/definitions/models.ProjectFile'
        type: array
    type: object
  models.ProjectFile:
    properties:
      content:
        type: string
      project_name:
        type: string
      relPath:
        type: string
      tag:
        type: string
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a SoftwareWarehouse API server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: SoftwareWarehouse Web API
  version: "1.0"
paths:
  /api/v1/binary/local:
    patch:
      consumes:
      - multipart/form-data
      parameters:
      - description: file
        in: formData
        name: file
        required: true
        type: file
      - description: projectName
        in: formData
        name: projectName
        required: true
        type: string
      - description: tag
        in: formData
        name: tag
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.LocalPatchBinaryResponse'
      summary: 本地上传二进制包
      tags:
      - binary
  /api/v1/source_code/local:
    patch:
      consumes:
      - multipart/form-data
      parameters:
      - description: file
        in: formData
        name: file
        required: true
        type: file
      - description: projectName
        in: formData
        name: projectName
        required: true
        type: string
      - description: tag
        in: formData
        name: tag
        required: true
        type: string
      - description: fileType
        in: formData
        name: fileType
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.LocalPatchSourceCodeResponse'
      summary: local上传源代码
      tags:
      - source_code
  /api/v1/source_code/search:
    get:
      parameters:
      - in: query
        name: code_type
        type: integer
      - in: query
        name: content
        type: string
      - in: query
        name: from
        type: integer
      - in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SearchSourceCodeResponse'
      summary: 模糊查询代码
      tags:
      - source_code
  /api/v1/test/error_handler:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: test error handler
      tags:
      - test
  /api/v1/test/ping_code_sim:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.PingCodeSimResponse'
      summary: ping code sim微服务测试
      tags:
      - test
swagger: "2.0"